Well... It scrapes it at least
Turns out the data was formated pretty badly in a lot of reviews and its
turning out to be a bigger pain that I intially thought it would.  The
code I originally wrote worked with only a subset of the reviews so I
got rid of it.  I'll get the formatting into a list of dictionaries done
tomorrow because its kinda late.